Move e2e test suites into subdirectories
