+++ functests/must-gather/must_gather_test.go
@@ -1,15 +1,31 @@
-package functests_test
+package must_gather_test
 
 import (
 	"fmt"
 	"os"
 	"os/exec"
+	"testing"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+
+	tests "github.com/openshift/ocs-operator/functests"
 )
 
-var _ = Describe("Must Gather", MustGatherTest)
+func TestTests(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "must-gather Test Suite")
+}
+
+var _ = BeforeSuite(func() {
+	tests.BeforeTestSuiteSetup()
+})
+
+var _ = AfterSuite(func() {
+	tests.AfterTestSuiteCleanup()
+})
+
+var _ = PDescribe("Must Gather", MustGatherTest)
 
 func MustGatherTest() {
 	BeforeEach(func() {
